07-Flask-JWT-Extended/requirements.txt
Flask-JWT_Extended
Flask-RESTful
Flask-SQLAlchemy